release!: 0.17.0
Squash of the unpublished Chatter work prepared after the three commits already present on public main. The body is the CHANGELOG's 0.17.0 section.

## [0.17.0] - 2026-08-30

### Changed

- **Reference-mode speaker identification now preserves typed lexical support.**
  `DonorMatchReport` retains the reference, donor, shared, and union token
  counts that derive each Jaccard score; its winner, evidence, and confidence
  margin are no longer independently constructible public fields. Thresholds
  are checked `ConfidenceThreshold` values, while confidence is explicitly
  `NoInformation`, `Finite`, or `Unbounded` instead of overloading `0.0` and
  infinity. Use the read-only report accessors in place of direct field access.
  `chatter speaker-id --write-match-report NEW.json` writes the accepted,
  low-confidence, structural-refusal, or input-refusal evidence without
  replacing an existing report.

- **`%wor` timing is now admitted through explicit typed evidence states.**
  `MainTier::wor_projection()` defines the shared positional membership policy;
  count binding, canonical-token corroboration, and complete positive interval
  assessment are separate states, so equal word counts alone cannot be treated
  as trustworthy timing. The impossible tier-level `WorTier::bullet` field is
  removed: timing evidence exists only when an actual `%wor` word carries a
  bullet. Callers of the former alignment and tier-bullet APIs must migrate to
  the projection, binding, corroboration, sequence-assessment, and
  `WorTier::timing_evidence()` APIs.

- **`rediarize` and `rediarize_content` require a `DiarizationTimeline`.**
  The windowed overlap algorithm needs turns ordered by start time, but the
  former `&[DiarizationTurn]` API let every library caller bypass that
  precondition and silently obtain a wrong winner. `DiarizationTimeline::new`
  owns the sorting transition, retains the longest-turn window bound, and
  keeps its ordered storage private. `TurnsFile` now exposes `source()` and
  `timeline()` accessors instead of independently public fields.

- **Tree-sitter 0.27.0** now drives the Rust parser, highlighting runtime,
  grammar crate, spec tooling, and grammar-generation CLI. The Node binding is
  independently current at 0.25.1. Generated parser artifacts and the complete
  parser/backend parity gates are regenerated and checked with this toolchain.

- **The vendored Rust lexer is regenerated with re2c 4.6.**
  `re2c-version.toml` is now the exact generator source of truth, and
  `just verify-vendored-lexer` refuses a different `re2rust` before comparing
  generated bytes. The 4.6 output differs from 4.5.1 only in its generator
  provenance header; upstream's 4.6 implementation change is Zig-only.

- **`Merged::report` is the only route from a merge to a file.** `into_file`
  and `file` are gone from `Merged`; `report(sink)` yields a `Reported`, which
  owns them. Serializing a merge without asking what it dropped is no longer
  writable, which is what two commands did until each was fixed by hand.
- **`chatter merge` and `chatter pipeline` now warn when a File 1 speaker is
  dropped.** A speaker outside `--retain` loses every utterance while keeping
  its `@Participants` row, so the output declares someone who says nothing.
  `AmbiguousSpeaker` does not catch it: that fires only when a code appears in
  both files. Both commands print the same warning, naming the speakers and how
  many utterances each lost, from one shared reporter. `chatter batch` drives
  the pipeline path, so the silent one was the path that runs whole corpora.

- **`merge_chat_files` returns the provenance of every merged utterance**, not
  only the merged file. It always knew this and threw it away: it walks each
  input in order building two lists, then stable-sorts the combination by
  `start_ms`. A consumer joining the output back to its inputs had to
  reconstruct the mapping by matching `(speaker, raw bullet)`, which is correct
  only while two facts hold that no caller can check, that the sort is stable
  and that inserted utterances are cloned unedited.

  The return type is now `Merged`, carrying the file, one `MergeOrigin` per
  output utterance in output order, one `ReferenceFate` per File 1 utterance
  and one `DonorFate` per File 2 utterance, each in its own input's order.
  Ordinals are `ReferenceIdx` / `DonorIdx`, separate types because two
  same-signature accessors over one index type answered confidently about the
  wrong file. `merge_chats`, the string wrapper, is **removed**: it had no non-test caller,
  and a `String` return cannot carry the provenance, so every consumer that
  wants the report has to work on parsed files anyway. `MergeError::Parse` goes
  with it, since `merge_chat_files` takes files that are already parsed.

  Prefer `Merged::utterances_with_origin()` to pairing the accessors by hand:
  zipping `origins()` against `file().lines` type-checks and is wrong by the
  number of header lines.

  `DonorFate` is a partition rather than a list of exclusions, so "this donor
  utterance is unaccounted for" is not expressible. An earlier form returned
  only the excluded ordinals and proved completeness with arithmetic, which
  balances just as well when every ordinal is shifted by the donor's header
  count.

  Both inputs are accounted for. `DonorFate` covers File 2; `ReferenceFate`
  covers File 1, where an utterance whose speaker is not retained is dropped and
  `AmbiguousSpeaker` does not catch it, because that fires only when a code
  appears in both files. A reference-only `MOT` with `retain = [CHI]` therefore
  passed every precondition, kept its `@Participants` row, and lost every
  utterance silently.

  `DonorFate::Inserted` carries `tiers_stripped`, because `strip_tiers` applies
  to the donor and only the donor: a bare `Inserted` claimed "carried over" for
  an utterance that was carried over AND edited.

  Still not a complete account of everything a merge omits, which is why the
  accessors are named `excluded_by_retain` and `dropped_not_retained` rather
  than `excluded` and `dropped`: donor headers other than `@ID` and `@Comment`
  are not carried.

### Fixed

- **`chatter pipeline --override-file` now refuses invalid override files.**
  Malformed TOML, unsupported schema versions, and read failures previously
  disappeared into "no override configured", silently triggering a fresh
  automatic speaker match. Explicit operator input now travels through the
  existing typed `OverrideFileError` exit path and no merged output is written.

- **`chatter rediarize` no longer counts overlapping turns from one track
  twice.** A track appearing in several turns is now measured by the UNION of
  its coverage: gaps remain gaps, while same-track overlaps count their shared
  interval once instead of manufacturing speaker time and distorting
  `--contested-at`. Cross-track overlap remains evidence for both simultaneous
  speakers, so `ownership.total_ms` is the sum of per-track union-held time and
  can exceed the utterance bullet's duration. The JSON shape is unchanged; its
  corrected measurement semantics are documented in the user guide.

- **`chatter fix --apply` can now repair E750 inside its recovered utterance.**
  Ordinary splice edits remain barred from parser-tainted regions. The E750
  catalog entry alone carries the typed state that it removes the delimiter
  whitespace responsible for that recovery, and the command still reparses and
  independently validates the resulting CHAT before writing it.

Red: compiler refusals, focused regression tests, parser snapshots, and the canonical just gate described in the release notes.

+++ crates/talkbank-parser/src/parser/chat_file_parser/header_dispatch/finder.rs
@@ -28,7 +28,7 @@ pub(super) fn find_header_node_in_tree(root: Node, index: usize) -> ParseResult<
     let mut found_count = 0;
 
     for i in 0..root.child_count() {
-        if let Some(child) = root.child(i as u32) {
+        if let Some(child) = root.child(i) {
             let kind = child.kind();
 
             // 1. Pre-begin header wrappers/subtypes
@@ -42,7 +42,7 @@ pub(super) fn find_header_node_in_tree(root: Node, index: usize) -> ParseResult<
                 } else {
                     // PRE_BEGIN_HEADER wrapper: look inside for header children
                     for j in 0..child.child_count() {
-                        if let Some(grandchild) = child.child(j as u32)
+                        if let Some(grandchild) = child.child(j)
                             && is_header(grandchild.kind())
                         {
                             if found_count == index {
@@ -58,7 +58,7 @@ pub(super) fn find_header_node_in_tree(root: Node, index: usize) -> ParseResult<
             // 2. LINE nodes: search inside for headers
             if kind == LINE {
                 for j in 0..child.child_count() {
-                    if let Some(grandchild) = child.child(j as u32) {
+                    if let Some(grandchild) = child.child(j) {
                         let gc_kind = grandchild.kind();
                         if gc_kind == HEADER {
                             // HEADER wrapper inside LINE: unwrap
