+++ crates/talkbank-parser/src/parser/tree_parsing/main_tier/content/word.rs
@@ -73,7 +73,7 @@ pub(crate) fn parse_word_content(
     }
 
     // Position 1+: optional whitespaces, replacement, base_annotations
-    while (idx as usize) < child_count {
+    while idx < child_count {
         if let Some(child) = node.child(idx) {
             match child.kind() {
                 WHITESPACES => {
